[java] Refactoring DevTools classes to match CDP specification structure
