+++ java/client/src/org/openqa/selenium/devtools/runtime/Runtime.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.devtools;
+package org.openqa.selenium.devtools.runtime;
 
 import java.util.Objects;
